backend/requirements-linux.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-recaptcha==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
idna==3.10
oauthlib==3.3.1
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
qrcode==8.2
reportlab==4.4.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.7
SpeechRecognition==3.14.3
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
